Refactor particles a lot
ParticleSystems now have a "Particles" object which is the actual
storage of the slices of things, and which creates its own animation
code for animations -- because animations need to be able to operate
on different layouts of things, so a thing which doesn't need all
the parameters can, well, have fewer parameters.

Also we drop some of the "initial state and context" stuff, which
does create some extra work -- you can't just ignore scale factors
when placing things. This needs to be revisited a bit more
carefully, but it means that we're no longer tracking, separately:
	* the screen X and Y, and theta, around which the particle's
	  local coordinate system is computed
	* the particle's current X and Y location, within that system
	* the particle's DX/DY
	* the particle's actual on-screen location, computed by
	  using the current X/Y location, multiplied by sin/cos of
	  the on-screen theta, etcetera

Now, you specify a coordinate in particle system space relative to
the particle system, and it translates those to screen coordinates,
and it also translates a provided dx/dy to screen coordinates,
*including* particle system rotation, and then just uses those values
in absolute terms thereafter.

So, when drawing:

< x, y := p.x0+((p.X*p.psCos-p.Y*p.psSin)*ps.Size), p.y0+((p.X*p.psSin+p.Y*p.psCos)*ps.Size)
> x, y := p.X, p.Y

The benchmark isn't really directly translatable anymore, but
some rough experiments suggest this is about 4x faster.